pyproject.toml
[project]
name = "upo-app-ai"
version = "0.1.0"
description = "Agente di Pianificazione Viaggi Intelligente"
requires-python = "==3.12.*"

# Qui ho messo alcune librerie utili utilizzate durante il corso.
# Se si nota che alcune non sono necessarie, si potrebbero rimuovere
# in modo da evitare conflitti o problemi futuri.
# Per ogni roba ho fatto un commento per evitare di dimenticarmi cosa fa chi.
# Inoltre ho messo una emoji per indicare se è raccomandato o meno.
dependencies = [
    # ✅ per i test
    "pytest",
    # ✅ per gestire variabili d'ambiente (generalmente API keys od opzioni)
    "dotenv",
    # 🟡 per fare scraping di pagine web
    #"bs4",
    # ✅ per fare una UI web semplice con input e output
    "gradio",
    # ✅ per la crittografia (necessaria per autenticazione Coinbase)
    "cryptography",
    # ❌ per l'elaborazione del linguaggio naturale in locale (https://huggingface.co/learn/llm-course/chapter1/3?fw=pt)
    #"transformers",
    # ❌ per fare chiamate a modelli indipendentemente dal modello specifico (astrae meglio rispetto a openai)
    #"langchain",
    # ✅ per costruire agenti (ovvero modelli che possono fare più cose tramite tool) https://github.com/agno-agi/agno
    # altamente consigliata dato che ha anche tools integrati per fare scraping, calcoli e molto altro
    # oltre a questa è necessario installare anche le librerie specifiche per i modelli che si vogliono usare
    "agno",
    # ✅ Modelli supportati e installati (aggiungere qui sotto quelli che si vogliono usare)
    "google-genai",
    "ollama",
    "openai",
    "anthropic",
    "google",
    "coinbase-advanced-py",
    "cryptocompare",
    "cdp-sdk",
    "python-binance"
]
